a_task_pool/mvp001/requirements.txt
pydantic>=2.5
pyyaml>=6.0
click>=8.1
eval_type_backport>=0.2  ; python_version < "3.10"
